backend/requirements.txt
Django==5.2.7
djangorestframework==3.16.1
crewai==1.1.
requests==2.32.5
pydantic==2.12.3
python-decouple==3.8
django-cors-headers==4.9.0
openai==1.109.1